requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.8
django-allauth==0.45.0
django-braces==1.14.0
django-cleanup==5.2.0
django-cpf==0.1.0
django-crispy-forms==1.13.0
idna==3.3
mysqlclient==2.0.3
oauthlib==3.1.1
pycodestyle==2.8.0
pycparser==2.20
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
